changed from double to float

+++ src/robot_arm.cpp
@@ -6,14 +6,14 @@
 /// @param angles Motor angles.
 /// @return The point.
 BLA::Matrix<3> forwardKinematics(const BLA::Matrix<3> angles) {
-    double theta1 = angles(0) * DEG_TO_RAD;
-    double theta2 = angles(1) * DEG_TO_RAD;
-    double theta3 = angles(2) * DEG_TO_RAD;
+    float theta1 = angles(0) * DEG_TO_RAD;
+    float theta2 = angles(1) * DEG_TO_RAD;
+    float theta3 = angles(2) * DEG_TO_RAD;
 
     // 2DOF arm position for the elbow and wrist in the XZ plane
     // local coordinates from second joint
-    double x_l = L2 * cos(theta2) + L3 * cos(theta2 + theta3);
-    double z_l = L2 * sin(theta2) + L3 * sin(theta2 + theta3);
+    float x_l = L2 * cos(theta2) + L3 * cos(theta2 + theta3);
+    float z_l = L2 * sin(theta2) + L3 * sin(theta2 + theta3);
 
     // apply rotation from the first joint (rotation along the z-axis)
     float x_w = cos(theta1) * x_l;
@@ -52,7 +52,7 @@ BLA::Matrix<3,3> estimateInitialJacobian() {
         P = forwardKinematics(temp_get_motor_angles());
         dP = P - last_P;
 
-        double delta_theta_i = delta_angles(i, i);
+        float delta_theta_i = delta_angles(i, i);
         J(0, i) = dP(0) / delta_theta_i;
         J(1, i) = dP(1) / delta_theta_i;
         J(2, i) = dP(2) / delta_theta_i;
@@ -71,20 +71,20 @@ BLA::Matrix<3,3> estimateInitialJacobian() {
 /// @return The updated Jacobian.
 BLA::Matrix<3,3> recomputeJacobian(const BLA::Matrix<3> angles) {
     
-    double theta1 = angles(0) * DEG_TO_RAD;
-    double theta2 = angles(1) * DEG_TO_RAD;
-    double theta3 = angles(2) * DEG_TO_RAD;
+    float theta1 = angles(0) * DEG_TO_RAD;
+    float theta2 = angles(1) * DEG_TO_RAD;
+    float theta3 = angles(2) * DEG_TO_RAD;
 
     // simplification
-    double c1 = cos(theta1), s1 = sin(theta1);
-    double c2 = cos(theta2), s2 = sin(theta2);
-    double c23 = cos(theta2 + theta3), s23 = sin(theta2 + theta3);
+    float c1 = cos(theta1), s1 = sin(theta1);
+    float c2 = cos(theta2), s2 = sin(theta2);
+    float c23 = cos(theta2 + theta3), s23 = sin(theta2 + theta3);
 
     BLA::Matrix<3,3> J;
 
     // intermediate terms; (x, z) relative to the 2DOF arm
-    double x_part = L2 * c2 + L3 * c23;
-    double z_part = L2 * s2 + L3 * s23;
+    float x_part = L2 * c2 + L3 * c23;
+    float z_part = L2 * s2 + L3 * s23;
 
     J = {
         -s1 * x_part, -c1 * z_part,       -c1 * L3 * s23,
